+++ packages/form-config/src/components/Antd/DraftFileList/utils.tsx
@@ -2,11 +2,7 @@ import React from 'react';
 import { Dictionary } from '@onaio/utils';
 import { ManifestFilesTypes } from '../../../ducks/manifestFiles';
 import { TableActions } from './TableActions';
-import { getFetchOptions, OpenSRPService } from '@opensrp/server-service';
-import { OPENSRP_MANIFEST_ENDPOINT, ERROR_OCCURRED } from '../../../constants';
-import { sendErrorNotification } from '@opensrp/notifications';
-import { removeManifestDraftFiles } from '../../../ducks/manifestDraftFiles';
-import { store } from '@onaio/redux-reducer-registry';
+import { getFetchOptions } from '@opensrp/server-service';
 import { formatDate } from '../../../helpers/utils';
 
 /**
@@ -68,47 +64,4 @@ export const getTableColumns = (
     },
   });
   return columns;
-};
-
-/**
- * Handle make release
- *
- * @param {ManifestFilesTypes[]} data draft files to make release for
- * @param {string} accessToken  Opensrp API access token
- * @param {string} opensrpBaseURL Opensrp API base URL
- * @param {Function} dispatch redux dispatch
- * @param {Function} removeDraftFiles redux action to remove draft files
- * @param {Function} setIfDoneHere set ifDoneHere form status
- * @param {Function} customFetchOptions custom opensrp API fetch options
- */
-export const makeRelease = (
-  data: ManifestFilesTypes[],
-  accessToken: string,
-  opensrpBaseURL: string,
-  dispatch: typeof store.dispatch,
-  removeDraftFiles: typeof removeManifestDraftFiles,
-  setIfDoneHere: (ifDoneHere: boolean) => void,
-  customFetchOptions?: typeof getFetchOptions
-) => {
-  const identifiers = data.map((form) => form.identifier);
-  const json = {
-    /* eslint-disable-next-line @typescript-eslint/camelcase */
-    forms_version: data[0].version,
-    identifiers,
-  };
-  const clientService = new OpenSRPService(
-    accessToken,
-    opensrpBaseURL,
-    OPENSRP_MANIFEST_ENDPOINT,
-    customFetchOptions
-  );
-  clientService
-    .create({ json: JSON.stringify(json) })
-    .then(() => {
-      dispatch(removeDraftFiles());
-      setIfDoneHere(true);
-    })
-    .catch((_: Error) => {
-      sendErrorNotification(ERROR_OCCURRED);
-    });
 };