+++ packages/keycloak-service/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.13 (2021-02-24)
+
+**Note:** Version bump only for package @opensrp/keycloak-service
+
+## [0.0.12](https://github.com/opensrp/web/compare/@opensrp/keycloak-service@0.0.11...@opensrp/keycloak-service@0.0.12) (2021-02-23)
+
+**Note:** Version bump only for package @opensrp/keycloak-service
+
+## [0.0.11](https://github.com/opensrp/web/compare/@opensrp/keycloak-service@0.0.10...@opensrp/keycloak-service@0.0.11) (2021-02-18)
+
+**Note:** Version bump only for package @opensrp/keycloak-service
+
+## 0.0.10 (2021-02-18)
+
+**Note:** Version bump only for package @opensrp/keycloak-service
+
 ## 0.0.8 (2020-11-12)
 
 **Note:** Version bump only for package @opensrp/keycloak-service